Renaming project to lho from messenger
The rename was important to transition from the experimental system to
the more definitive project.
