+++ pom.xml
@@ -3,14 +3,14 @@
 
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.ronel.tutorial</groupId>
-    <artifactId>messenger</artifactId>
+    <groupId>com.draag.services</groupId>
+    <artifactId>lho</artifactId>
     <packaging>war</packaging>
     <version>1.0-SNAPSHOT</version>
     <name>messenger</name>
 
     <build>
-        <finalName>messenger</finalName>
+        <finalName>lho</finalName>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
