+++ src/main/java/com/draag/services/MyResource.java
@@ -1,4 +1,4 @@
-package org.ronel.tutorial.messenger;
+package com.draag.services;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
