+++ src/main/java/com/draag/services/service/MessageService.java
@@ -1,12 +1,12 @@
-package org.ronel.tutorial.messenger.service;
+package com.draag.services.service;
 
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
 import java.util.Map;
 
-import org.ronel.tutorial.messenger.DatabaseClass;
-import org.ronel.tutorial.messenger.model.Message;
+import com.draag.services.DatabaseClass;
+import com.draag.services.model.Message;
 
 public class MessageService {
